Added function for external content
External content should now go in the dedicated folders, e.g. fonts in the font folder, images in the images folder, icalendars in the icalendar folder. This poses a problem where the filename is known but not the full path.

This function returns the full path of a given filename, e.g. 'pikachu.png' will return the full path of the image e.g. '/home/.../external/images/pikachu.png'.

By checking the file extension, it automatically knows which folder to search for.